✨ feat(borrowedLog): 借用日志
